new loading screen & leave game
A loading screen that can be trigger with the toast service is implemented. It was needed since some server side function of appwrite take some time to respond, it's to prevent the user to panic !

Implemented the loading screen into
- JoinGame
- CreateGame
- LeaveGame
- DeleteGame
- CreatePlayer

Implemented the function to delete a player from a game

Implemented the leaveGame function. When leaving a game, the user will delete their player within the game